python深度学习/Keras应用/第八章-深度学习用于计算机视觉/requirements.txt
# 猫狗图像分类项目依赖
# 《Python深度学习》第5章配套代码

tensorflow>=2.10
numpy>=1.21
matplotlib>=3.5
Pillow>=9.0